requirements.txt
# Core ML
torch>=2.1
torchvision>=0.16
numpy>=1.24
scipy>=1.11

# Object Detection
# YOLOX via official repo (not mmdet — simpler deps)
# Install separately: pip install git+https://github.com/Megvii-BaseDetection/YOLOX.git
pycocotools>=2.0

# VLM
transformers>=4.40
accelerate>=0.27
bitsandbytes>=0.43
qwen-vl-utils>=0.0.8

# XAI
captum>=0.7
scikit-image>=0.22    # SLIC superpixels for LIME/D-CLOSE
lime>=0.2             # LIME base library

# Evaluation & Utilities
opencv-python>=4.8
Pillow>=10.0
matplotlib>=3.8
seaborn>=0.13
pandas>=2.1
scikit-learn>=1.3
pyyaml>=6.0
json-repair>=0.28
tqdm>=4.66

# Logging & Reproducibility
wandb>=0.16           # Optional: experiment tracking
tensorboard>=2.14     # Optional: alternative to wandb

# Testing
pytest>=7.4
pytest-cov>=4.1

